requirements-local.txt
alembic==1.18.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
freez==0.0.0
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.20.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==26.0
pillow==12.1.0
# PostgreSQL driver removed - not needed for local development with SQLite
# psycopg2-binary==2.9.11
# SQL Server driver removed - causes build errors on Windows without C++ tools
# pyodbc==5.2.0
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.15.0
Werkzeug==3.0.1
wrapt==2.0.1

# Note: This file is for LOCAL DEVELOPMENT on Windows
# Docker uses the full requirements.txt which includes database drivers
# Use SQLite for local dev: DATABASE_URL=sqlite:///instance/ecommerce_dev.db
